+++ src/lib/adapters/WebDav.ts
@@ -92,7 +92,7 @@ export default class WebDavAdapter extends CachingAdapter {
       if (res.headers['Last-Modified']) {
         const date = new Date(res.headers['Last-Modified'])
         const dateLocked = date.valueOf()
-        if (dateLocked > Date.now() - LOCK_TIMEOUT) {
+        if (dateLocked + LOCK_TIMEOUT > Date.now()) {
           throw new ResourceLockedError()
         }
       } else {
@@ -102,6 +102,7 @@ export default class WebDavAdapter extends CachingAdapter {
 
     if (res.status === 200) {
       // continue anyway
+      this.locked = true
     } else if (res.status === 404) {
       await this.setLock()
     } else {
@@ -120,7 +121,15 @@ export default class WebDavAdapter extends CachingAdapter {
       'text/html',
       '<html><body>I am a lock file</body></html>'
     )
-    await this.lockingPromise
+    try {
+      await this.lockingPromise
+    } catch (e) {
+      if (e instanceof HttpError && e.status === 423) {
+        this.locked = false
+        throw new ResourceLockedError()
+      }
+      throw e
+    }
     this.locked = true
   }
 
@@ -238,7 +247,13 @@ export default class WebDavAdapter extends CachingAdapter {
 
     if (Capacitor.getPlatform() === 'web') {
       const browser = (await import('../browser-api')).default
-      if (!(await browser.permissions.contains({ origins: [this.server.url + '/'] }))) {
+      let hasPermissions
+      try {
+        hasPermissions = await browser.permissions.contains({ origins: [this.server.url + '/'] })
+      } catch (e) {
+        console.warn(e)
+      }
+      if (!hasPermissions) {
         throw new MissingPermissionsError()
       }
     }
@@ -247,8 +262,12 @@ export default class WebDavAdapter extends CachingAdapter {
       throw new SlashError()
     }
 
+    if (this.lockingInterval) {
+      clearInterval(this.lockingInterval)
+    }
     if (needLock) {
       await this.obtainLock()
+      this.lockingInterval = setInterval(() => this.setLock(), LOCK_INTERVAL) // Set lock every minute
     }
 
     const resp = await this.pullFromServer()
@@ -258,8 +277,6 @@ export default class WebDavAdapter extends CachingAdapter {
         throw new HttpError(resp.status, 'GET')
       }
     }
-
-    this.lockingInterval = setInterval(() => this.setLock(), LOCK_INTERVAL) // Set lock every minute
 
     this.initialTreeHash = await this.bookmarksCache.hash(true)