+++ webpack.prod.js
@@ -8,9 +8,12 @@ const webpack = require('webpack')
 module.exports = merge(common, {
   mode: 'production',
   devtool: 'source-map',
+  optimization: {
+    splitChunks: { chunks: 'all' },
+  },
   plugins: [
     new webpack.DefinePlugin({
-      DEBUG: JSON.stringify(!process.env['CI'])
+      DEBUG: JSON.stringify(false)
     }),
     sentryWebpackPlugin({
       authToken: process.env.SENTRY_AUTH_TOKEN,